Rename user->member
API 설계대로 user에서 member로 전체 변경

이슈 없음
